Fix 64-bit variant index corruption bug and add test case
- Change add_scoped_variant and create_scoped_variant return types from unsigned to int32_t
- Change create_permanent_variant parameter and return type from unsigned to int32_t
- Fix variant index handling to preserve negative values for permanent variants
- Add explicit int32_t casts when passing variant indices to get_scoped_variant
- Fixes corrupted scoped variant index errors in 64-bit builds
- Add test_variant_index_64bit_corruption test case to verify fix
- Clean up MinGW cross-compilation files (restore build_mingw.sh, remove mingw_toolchain.cmake)

+++ src/guest_variant.cpp
@@ -45,7 +45,7 @@ Variant GuestVariant::toVariant(const Sandbox &emu) const {
 		}
 
 		default:
-			if (std::optional<const Variant *> v = emu.get_scoped_variant(this->v.i)) {
+			if (std::optional<const Variant *> v = emu.get_scoped_variant(static_cast<int32_t>(this->v.i))) {
 				const Variant *var = *v;
 				return *var;
 			} else {
@@ -58,7 +58,7 @@ Variant GuestVariant::toVariant(const Sandbox &emu) const {
 }
 
 const Variant *GuestVariant::toVariantPtr(const Sandbox &emu) const {
-	if (std::optional<const Variant *> v = emu.get_scoped_variant(this->v.i))
+	if (std::optional<const Variant *> v = emu.get_scoped_variant(static_cast<int32_t>(this->v.i)))
 		return v.value();
 
 	char buffer[128];
@@ -217,7 +217,7 @@ void GuestVariant::create(Sandbox &emu, Variant &&value) {
 
 		default: {
 			// Store the variant in the current state
-			unsigned int idx = emu.create_scoped_variant(std::move(value));
+			int32_t idx = emu.create_scoped_variant(std::move(value));
 			this->v.i = idx;
 		}
 	}
